pyproject.toml
[project]
name = "kraw"
version = "0.1.0"
description = "k8s discovery across AWS, Azure etc."
authors = [
    {name = "Kodanda Rama",email = "kodrama@google.com"},
    {name = "Surabhi Sharma",email = "surabhiisharma@google.com"},
    {name = "San Srinivasan",email = "sansrinivasan@google.com"},

]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-identity>=1.24.0",
    "azure-mgmt-containerservice>=39.0.0",
    "azure-mgmt-resource>=24.0.0",
    "boto3>=1.40.8",
    "kubernetes>=33.1.0",
    "pandas>=2.3.1",
    "rich>=14.1.0",
    "typer[all]>=0.16.0",
]

[dependency-groups]
dev = [
    "pyinstaller>=6.15.0",
]
lint = [
    "ruff>=0.12.8",
]
